+++ CHANGES.md
@@ -11,6 +11,9 @@ drop
 core
 ----
 
+- [ADD] Add `println!()`, `eprintln!()` and `dbg!()` for WebAssembly build
+- [CHANGE] Remove `log` crate from dependencies
+- [CHANGE] Remove `System::console_log`
 - [CHANGE] Remove `Event::Terminating`
 - [CHANGE] Remove `System::state_XXX()` methods in favor of `Game::command()` and `Game::query()`
 - [UPDATE] Implement `Default` for `Sprite`